Add "FinalizedBlock" tag and related methods
Introduced a new "FinalizedBlock" tag in the rpc/types.go file. Also, added the corresponding IsFinalized() method and updated the existing method to recognize this tag. Added a new FinalizeBlockNum() function to the rpc client in the rpc/rpc.go file for handling the eth_blockNumber request related to finalized blocks.